# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LMS API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterRequest {
  email: string;
  /** @minLength 8 */
  password: string;
  name: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface UserResponse {
  id: number;
  email: string;
  name: string;
  createdAt: string;
}

export interface AuthResponse {
  accessToken: string;
  user: UserResponse;
}

export interface RefreshResponse {
  accessToken: string;
}

export interface SubjectSummary {
  id: number;
  title: string;
  slug: string;
  description?: string | null;
  isPublished: boolean;
  createdAt: string;
  totalVideos: number;
  totalSections: number;
}

export interface SubjectsListResponse {
  subjects: SubjectSummary[];
  total: number;
  page: number;
  limit: number;
}

export interface SubjectDetailResponse {
  id: number;
  title: string;
  slug: string;
  description?: string | null;
  isPublished: boolean;
  createdAt: string;
  totalVideos: number;
  totalSections: number;
  isEnrolled: boolean;
}

export interface TreeVideo {
  id: number;
  title: string;
  orderIndex: number;
  durationSeconds?: number | null;
  locked: boolean;
  isCompleted: boolean;
  lastPositionSeconds?: number | null;
}

export interface TreeSection {
  id: number;
  title: string;
  orderIndex: number;
  videos: TreeVideo[];
}

export interface SubjectTreeResponse {
  id: number;
  title: string;
  sections: TreeSection[];
}

export interface EnrollmentStatusResponse {
  isEnrolled: boolean;
  enrolledAt?: string | null;
}

export interface VideoDetailResponse {
  id: number;
  title: string;
  description?: string | null;
  youtubeUrl: string;
  orderIndex: number;
  durationSeconds?: number | null;
  subjectId: number;
  sectionId: number;
  sectionTitle: string;
  previousVideoId?: number | null;
  nextVideoId?: number | null;
  locked: boolean;
  unlockReason?: string | null;
}

export interface VideoProgressResponse {
  videoId: number;
  lastPositionSeconds: number;
  isCompleted: boolean;
  completedAt?: string | null;
}

export interface UpdateProgressRequest {
  lastPositionSeconds: number;
  isCompleted?: boolean;
}

export interface SubjectProgressResponse {
  subjectId: number;
  totalVideos: number;
  completedVideos: number;
  percentComplete: number;
  lastVideoId?: number | null;
  lastPositionSeconds?: number | null;
}

export type ListSubjectsParams = {
  page?: number;
  limit?: number;
  search?: string;
};