# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LMS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResponse"
        "401":
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /subjects/{subjectId}:
    get:
      operationId: getSubject
      tags: [subjects]
      summary: Get subject metadata
      parameters:
        - in: path
          name: subjectId
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Subject details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubjectDetailResponse"
        "404":
          description: Subject not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /videos/{videoId}:
    get:
      operationId: getVideo
      tags: [videos]
      summary: Get video details with lock status
      parameters:
        - in: path
          name: videoId
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Video details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoDetailResponse"
        "403":
          description: Not enrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Video not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /progress/videos/{videoId}:
    get:
      operationId: getVideoProgress
      tags: [progress]
      summary: Get video progress
      parameters:
        - in: path
          name: videoId
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Video progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoProgressResponse"

    post:
      operationId: updateVideoProgress
      tags: [progress]
      summary: Update video progress
      parameters:
        - in: path
          name: videoId
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProgressRequest"
      responses:
        "200":
          description: Progress updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoProgressResponse"
  /progress/subjects/{subjectId}:
    get:
      operationId: getSubjectProgress
      tags: [progress]
      summary: Get aggregated subject progress
      parameters:
        - in: path
          name: subjectId
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Subject progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubjectProgressResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
      required:
        - email
        - password
        - name
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        user:
          $ref: "#/components/schemas/UserResponse"
      required:
        - accessToken
        - user
    RefreshResponse:
      type: object
      properties:
        accessToken:
          type: string
      required:
        - accessToken
    UserResponse:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - createdAt
...
    SubjectDetailResponse:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        isPublished:
          type: boolean
        createdAt:
          type: string
          format: date-time
        totalVideos:
          type: integer
        totalSections:
          type: integer
        isEnrolled:
          type: boolean
      required:
        - id
        - title
        - slug
        - isPublished
        - createdAt
        - totalVideos
        - totalSections
        - isEnrolled
...
    VideoDetailResponse:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        youtubeUrl:
          type: string
        orderIndex:
          type: integer
        durationSeconds:
          type: integer
          nullable: true
        subjectId:
          type: integer
        sectionId:
          type: integer
        sectionTitle:
          type: string
        previousVideoId:
          type: integer
          nullable: true
        nextVideoId:
          type: integer
          nullable: true
        locked:
          type: boolean
        unlockReason:
          type: string
          nullable: true
      required:
        - id
        - title
        - youtubeUrl
        - orderIndex
        - subjectId
        - sectionId
        - sectionTitle
        - locked
    VideoProgressResponse:
      type: object
      properties:
        videoId:
          type: integer
        lastPositionSeconds:
          type: integer
        isCompleted:
          type: boolean
        completedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - videoId
        - lastPositionSeconds
        - isCompleted
    UpdateProgressRequest:
      type: object
      properties:
        lastPositionSeconds:
          type: integer
        isCompleted:
          type: boolean
      required:
        - lastPositionSeconds
    SubjectProgressResponse:
      type: object
      properties:
        subjectId:
          type: integer
        totalVideos:
          type: integer
        completedVideos:
          type: integer
        percentComplete:
          type: number
        lastVideoId:
          type: integer
          nullable: true
        lastPositionSeconds:
          type: integer
          nullable: true
      required:
        - subjectId
        - totalVideos
        - completedVideos
        - percentComplete