# lib/api-zod/src/generated/types/subjectDetailResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LMS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SubjectDetailResponse {
  id: number;
  title: string;
  slug: string;
  description?: string | null;
  isPublished: boolean;
  createdAt: Date;
  totalVideos: number;
  totalSections: number;
  isEnrolled: boolean;
}