# lib/api-zod/src/generated/types/videoProgressResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LMS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VideoProgressResponse {
  videoId: number;
  lastPositionSeconds: number;
  isCompleted: boolean;
  completedAt?: Date | null;
}